Create initial structure
Create a new repo for this project on GitHub.com and call it
odin-recipes and choose the public option.
Clone that repository onto your local machine, inside the repos folder.
Set up a README.md file.
Within the odin-recipes directory, create an index.html file.
Fill it out with the usual boilerplate HTML
and add a title “Odin Recipes” to the head as well as
an h1 heading “Odin Recipes” to the body.