Add recipe page
Create a new directory within the odin-recipes directory and name it
recipes. Create a new HTML file within the recipes directory
and name it after the recipe it will contain (ground_beef_goulash.html).
Include an h1 heading with the recipe’s name as its content
(Ground Beef Goulash). Also add a title “Ground Beef Goulash Recipe” to
the head. Back in the index.html file, add a link to the recipe page
you just created.

Under the <h1>Odin Recipes</h1> heading, write out the link like so:
<a href="recipes/ground_beef_goulash.html">Ground Beef Goulash</a>.
The text of the link is again the recipe name.